Refactored the code to be cleaner with autos and teleops
