+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/BallTeleOp.java
@@ -1,32 +1,17 @@
 package org.firstinspires.ftc.teamcode.opmodes;
 
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
 import org.firstinspires.ftc.teamcode.control.InputControlManager;
-import org.firstinspires.ftc.teamcode.control.Statics;
-import org.firstinspires.ftc.teamcode.io.output.Log;
-import org.firstinspires.ftc.teamcode.io.output.LoggerThread;
 
 @TeleOp(name = "bla Teleop", group = "tele")
-public class BallTeleOp extends OpMode {
+public class BallTeleOp extends BaseTeleOp {
 
     private InputControlManager input;
 
     @Override
     public void init() {
-        Log.self = null;
-        Statics.set(telemetry, gamepad1, gamepad2, hardwareMap);
-        input = InputControlManager.get();
-    }
-
-    @Override
-    public void loop() {
-        input.update();
-    }
-
-    @Override
-    public void stop() {
-        LoggerThread.RUN = false;
+        super.init();
+        InputControlManager.USE_TIMER = false;
     }
 }
